jeSLIME 0.1.2--
   * Cell classes now have a "feed" method (unsupported in SimpleCell),
which informs cells of a direct benefit. The effect on fitness depends
on the class.

  * FissionCell (divide cuts its biomass in half, reports
isDivisible=true when it reaches a critical biomass). This means that
growth processes with FissionCells take place on a threshold basis:
since only divisible cells are polled for division events, only  those
FissionCells that have exceeded their biomass threshold will be
considered for division.

  * UniformBiomassGrowth process -- activates "feed" ability for every
cell by a defined delta.

  * Divisible sites index in lattice should change whenever cell state
might. Updated tests to reflect this.

  * "Cell descriptors" now used by a "Cell factory" to define properties
of cells to be created (Scatter process, etc) --> more flexible project
spec. "sequential" mode for state id allows unique ids to be generated.

  * Scatter now uses cell descriptors.