Upgraded Eco to Java 8. All 715 tests pass.
The purpose of the upgrade was to allow the use of lambdas, which promise to greatly improve code elegance and simplicity. They also may improve performance, since lambdas have automatic support for concurrency in collection traversal.

The upgrade broke several tests. Image output tests produced binary files that were different (presumably, Java 8 has some improvements to the graphics library). Another two tests that had the order of unordered collections hard-coded in also broke.

One of these--the test for OccupiedNeighborSwap--turned out to be passing for the wrong reasons (silent failure), so I rewrote the test to do what it was actually thought to do. It does not appear that there was anything wrong with the tested method, just the test itself.